+++ src/components/Button/Button.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import styles from './Button.module.scss'
 
 const Button = ({width, children, onClick}) => (
@@ -10,5 +11,16 @@ const Button = ({width, children, onClick}) => (
       {children}
     </div>
 )
+
+Button.propTypes = {
+    width: PropTypes.number,
+    children: PropTypes.any,
+    onClick: PropTypes.func.isRequired,
+}
+
+Button.defaultProps = {
+    width: null,
+    children: null,
+}
 
 export default Button
