+++ src/components/Button/Button.module.scss
@@ -1,9 +1,8 @@
-@import "../../../variables";
+@import "../../variables";
 
 .button {
   opacity: 1;
   padding: 5px;
-  //box-shadow: 0 0 0 1px white;
   border: 1px solid white;
   font-size: 1rem;
   color: white;
