+++ src/components/Modal/Modal.jsx
@@ -1,9 +1,10 @@
 import ReactDOM from 'react-dom'
 import React from 'react'
-import Trailer from '../../Trailer/Trailer'
+import Trailer from '../Trailer/Trailer'
 
 const Modal = ({callback, movieId}) => ReactDOM.createPortal(
     <Trailer callback={callback} movieId={movieId}/>,
-    document.querySelector('#app'))
+    document.querySelector('#app'),
+)
 
 export default Modal
