+++ src/components/Rating/Rating.module.scss
